requirements.txt
streamlit>=1.25
streamlit-mic-recorder>=0.0.6
faster-whisper>=0.6.0
jaconv>=0.3.5
edge-tts>=6.1.0
python-dotenv>=1.0.0
requests>=2.28.0
pykakasi>=2.2.1
